Bump LiteCommands dependency.
